// next-app/src/common/interface.ts
export interface AttributeValuePayload {
  value: string;
  description?: string;
  status: boolean;
  attribute_id: number;
};